# svertexer: adaptive mass windows