--- a/Detectors/Vertexing/include/DetectorsVertexing/SVertexHypothesis.h
+++ b/Detectors/Vertexing/include/DetectorsVertexing/SVertexHypothesis.h
@@ -33,15 +33,15 @@ class SVertexHypothesis
   enum PIDParams { SigmaM,  // sigma of mass res at 0 pt
                    NSigmaM, // number of sigmas of mass res
                    MarginM, // additive safety margin in mass cut
-                   CPt,   // pT dependence of mass resolution parameterized as mSigma*(1+mC1*pt);
+                   CPt,     // pT dependence of mass resolution parameterized as mSigma*(1+mC1*pt);
                    CPt1,
                    CPt2,
-                   CPt3 };   // pT dependence of mass resolution of Cascade parameterized as CPt+CPt1*pt +CPt2*TMath::Exp(-CPt3*pt);
+                   CPt3 }; // pT dependence of mass resolution of Cascade parameterized as CPt+CPt1*pt +CPt2*TMath::Exp(-CPt3*pt);
 
   static constexpr int NPIDParams = 7;
 
-  void set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float bz = 0.f);
-  void set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz = 0.f);
+  void set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float cpt1, float cpt2, float cpt3, float bz = 0.f, float maxSigma = 0.01);
+  void set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz = 0.f, float maxSigma = 0.01);
 
   float getMassV0Hyp() const { return PID::getMass(mPIDV0); }
   float getMassPosProng() const { return PID::getMass(mPIDPosProng); }
@@ -66,15 +66,21 @@ class SVertexHypothesis
     return std::abs(mass - getMassV0Hyp()) < getMargin(pt);
   }
 
-  float getSigmaCascade(float pt) const { return mPars[CPt]+mPars[CPt1]*pt +mPars[CPt2]*std::exp(-mPars[CPt3]*pt); }
+  float getSigmaV0Cascade(float pt) const { return mPars[CPt] + mPars[CPt1] * pt + mPars[CPt2] * std::exp(-mPars[CPt3] * pt); }
   float getSigma(float pt) const { return mPars[SigmaM] * (1.f + mPars[CPt] * pt); }
   float getMargin(float pt) const
   {
-    if( mPIDV0 == PID::XiMinus || mPIDV0 == PID::OmegaMinus) {
-      if( mPars[NSigmaM] * getSigmaCascade(pt) + mPars[MarginM] > 0.006 ) return mPars[NSigmaM] * 0.006;
-      return mPars[NSigmaM] * getSigmaCascade(pt) + mPars[MarginM];
-    }else if( mPIDV0 == PID::K0 || mPIDV0 == PID::Lambda ) return mPars[NSigmaM] * getSigmaCascade(pt) + mPars[MarginM];
-    else return mPars[NSigmaM] * getSigma(pt) + mPars[MarginM];
+    if (mPIDV0 == PID::XiMinus || mPIDV0 == PID::OmegaMinus) { // case for cascades, antiparticles included
+      if (getSigmaV0Cascade(pt) > maxSigma) {                  // insuring that at low pt one gets reasonable width as the parametrisation function may explode to unphysical values
+        return mPars[NSigmaM] * maxSigma + mPars[MarginM];
+      } else {
+        return mPars[NSigmaM] * getSigmaV0Cascade(pt) + mPars[MarginM];
+      }
+    } else if (mPIDV0 == PID::K0 || mPIDV0 == PID::Lambda) { // case for V0s, AntiLambda is included in PID::Lambda
+      return mPars[NSigmaM] * getSigmaV0Cascade(pt) + mPars[MarginM];
+    } else {
+      return mPars[NSigmaM] * getSigma(pt) + mPars[MarginM]; // case for HyperTriton and Hyperhydrog4
+    }
   }
 
  private:
@@ -87,6 +93,7 @@ class SVertexHypothesis
 
  public: // to be deleted
   std::array<float, NPIDParams> mPars{};
+  float maxSigma;
 
   ClassDefNV(SVertexHypothesis, 2);
 };