--- a/Detectors/Vertexing/include/DetectorsVertexing/SVertexHypothesis.h
+++ b/Detectors/Vertexing/include/DetectorsVertexing/SVertexHypothesis.h
@@ -33,14 +33,16 @@ class SVertexHypothesis
   enum PIDParams { SigmaM,  // sigma of mass res at 0 pt
                    NSigmaM, // number of sigmas of mass res
                    MarginM, // additive safety margin in mass cut
+                   NSigmaTightM, // number of sigmas of mass res when doing tight cut around mass (V0s used in cascades)
+                   MarginTightM, // additive safety margin in mass cut when doing tight cut around mass (V0s used in cascades)
                    CPt,     // pT dependence of mass resolution parameterized as mSigma*(1+mC1*pt);
                    CPt1,
                    CPt2,
                    CPt3 }; // pT dependence of mass resolution of Cascade parameterized as CPt+CPt1*pt +CPt2*TMath::Exp(-CPt3*pt);
 
-  static constexpr int NPIDParams = 7;
+  static constexpr int NPIDParams = 9;
 
-  void set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float cpt1, float cpt2, float cpt3, float bz = 0.f, float maxSigma = 0.01);
+  void set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float nSigTight, float marginTight, float cpt, float cpt1, float cpt2, float cpt3, float bz = 0.f, float maxSigma = 0.01);
   void set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz = 0.f, float maxSigma = 0.01);
 
   float getMassV0Hyp() const { return PID::getMass(mPIDV0); }
@@ -60,28 +62,43 @@ class SVertexHypothesis
   { // check if given mass and pt is matching to hypothesis
     return check(calcMass(p2Pos, p2Neg, p2V0), ptV0);
   }
-
   bool check(float mass, float pt) const
   { // check if given mass and pt is matching to hypothesis
     return std::abs(mass - getMassV0Hyp()) < getMargin(pt);
   }
 
+  bool checkTight(float p2Pos, float p2Neg, float p2V0, float ptV0) const
+  { // check if given mass and pt is matching to hypothesis
+    return checkTight(calcMass(p2Pos, p2Neg, p2V0), ptV0);
+  }
+  bool checkTight(float mass, float pt) const
+  { // check if given mass and pt is matching to hypothesis
+    return std::abs(mass - getMassV0Hyp()) < getMarginTight(pt);
+  }
+
   float getSigmaV0Cascade(float pt) const { return mPars[CPt] + mPars[CPt1] * pt + mPars[CPt2] * std::exp(-mPars[CPt3] * pt); }
   float getSigma(float pt) const { return mPars[SigmaM] * (1.f + mPars[CPt] * pt); }
-  float getMargin(float pt) const
+  float getMargin(float pt, bool tight = false) const
   {
+    int idxNsigma = NSigmaM;
+    int idxMargin = MarginM;
+    if(tight){ // move to indices for tight variables in case asked to do so (tighter peak cuts for decay chains)
+      idxNsigma = NSigmaTightM;
+      idxMargin = MarginTightM;  
+    }
     if (mPIDV0 == PID::XiMinus || mPIDV0 == PID::OmegaMinus) { // case for cascades, antiparticles included
       if (getSigmaV0Cascade(pt) > maxSigma) {                  // insuring that at low pt one gets reasonable width as the parametrisation function may explode to unphysical values
-        return mPars[NSigmaM] * maxSigma + mPars[MarginM];
+        return mPars[idxNsigma] * maxSigma + mPars[idxMargin];
       } else {
-        return mPars[NSigmaM] * getSigmaV0Cascade(pt) + mPars[MarginM];
+        return mPars[idxNsigma] * getSigmaV0Cascade(pt) + mPars[idxMargin];
       }
     } else if (mPIDV0 == PID::K0 || mPIDV0 == PID::Lambda) { // case for V0s, AntiLambda is included in PID::Lambda
-      return mPars[NSigmaM] * getSigmaV0Cascade(pt) + mPars[MarginM];
+      return mPars[idxNsigma] * getSigmaV0Cascade(pt) + mPars[idxMargin];
     } else {
-      return mPars[NSigmaM] * getSigma(pt) + mPars[MarginM]; // case for HyperTriton and Hyperhydrog4
+      return mPars[idxNsigma] * getSigma(pt) + mPars[idxMargin]; // case for HyperTriton and Hyperhydrog4
     }
   }
+  float getMarginTight(float pt) const { return getMargin(pt, true); }
 
  private:
   float getMass2PosProng() const { return PID::getMass2(mPIDPosProng); }