--- a/Detectors/Vertexing/src/SVertexHypothesis.cxx
+++ b/Detectors/Vertexing/src/SVertexHypothesis.cxx
@@ -17,7 +17,7 @@
 
 using namespace o2::vertexing;
 
-void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float bz)
+void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float cpt1, float cpt2, float cpt3, float bz, float maxSigma)
 {
   mPIDV0 = v0;
   mPIDPosProng = ppos;
@@ -25,13 +25,19 @@ void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, f
   mPars[SigmaM] = sig;
   mPars[NSigmaM] = nSig;
   mPars[MarginM] = margin;
+  mPars[CPt] = cpt;
+  mPars[CPt1] = cpt1;
+  mPars[CPt2] = cpt2;
+  mPars[CPt3] = cpt3;
+  maxSigma = maxSigma;
   float absBz{std::abs(bz)};
-  mPars[CPt] = absBz > 1e-3 ? cpt * 5.0066791 / absBz : 0.; // assume that pT dependent sigma is linear with B
+  if (cpt3 < 1)
+    mPars[CPt] = absBz > 1e-3 ? cpt * 5.0066791 / absBz : 0.; // assume that pT dependent sigma is linear with B; case for HyperTriton and Hyperhydrog4
 }
 
-void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz)
+void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz, float maxSigma)
 {
-  set(v0, ppos, pneg, pars[SigmaM], pars[NSigmaM], pars[MarginM], pars[CPt], bz);
+  set(v0, ppos, pneg, pars[SigmaM], pars[NSigmaM], pars[MarginM], pars[CPt], pars[CPt1], pars[CPt2], pars[CPt3], bz, maxSigma);
 }
 
 void SVertex3Hypothesis::set(PID v0, PID ppos, PID pneg, PID pbach, float sig, float nSig, float margin, float cpt, float bz)