--- a/Detectors/Vertexing/src/SVertexHypothesis.cxx
+++ b/Detectors/Vertexing/src/SVertexHypothesis.cxx
@@ -17,7 +17,7 @@
 
 using namespace o2::vertexing;
 
-void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float cpt, float cpt1, float cpt2, float cpt3, float bz, float maxSigma)
+void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, float margin, float nSigTight, float marginTight, float cpt, float cpt1, float cpt2, float cpt3, float bz, float maxSigma)
 {
   mPIDV0 = v0;
   mPIDPosProng = ppos;
@@ -25,6 +25,8 @@ void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, f
   mPars[SigmaM] = sig;
   mPars[NSigmaM] = nSig;
   mPars[MarginM] = margin;
+  mPars[NSigmaTightM] = nSigTight;
+  mPars[MarginTightM] = marginTight;
   mPars[CPt] = cpt;
   mPars[CPt1] = cpt1;
   mPars[CPt2] = cpt2;
@@ -37,7 +39,7 @@ void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, float sig, float nSig, f
 
 void SVertexHypothesis::set(PID v0, PID ppos, PID pneg, const float pars[NPIDParams], float bz, float maxSigma)
 {
-  set(v0, ppos, pneg, pars[SigmaM], pars[NSigmaM], pars[MarginM], pars[CPt], pars[CPt1], pars[CPt2], pars[CPt3], bz, maxSigma);
+  set(v0, ppos, pneg, pars[SigmaM], pars[NSigmaM], pars[MarginM], pars[NSigmaTightM], pars[MarginTightM], pars[CPt], pars[CPt1], pars[CPt2], pars[CPt3], bz, maxSigma);
 }
 
 void SVertex3Hypothesis::set(PID v0, PID ppos, PID pneg, PID pbach, float sig, float nSig, float margin, float cpt, float bz)