This is Lucia's work to parametrize V0 and cascade masses in a more accurate way. Some final testing is still ongoing. Note also that, as before, parameters are hardcoded; it may pay off to make them settable so a minor commit to do that might still come later today. More soon...