@shahor02 2023-10-26
Hi @ddobrigk 
@davidrohr has compared the performance of this PR with the current mass validation version on 25kHz pbpb data. 
To me the reduction of cascades is too strong:

Current selection
```
shahoian@alicrs02:~/Downloads$ grep 'strange tracks' sv0.log 
[2656164:secondary-vertexing]: [13:44:55][INFO] Found 380168 V0s, 3268892 cascades, 121875 3-body decays, 0 strange tracks. Timing: CPU: 278.14 Real: 11.4833 s
[2656164:secondary-vertexing]: [13:47:17][INFO] Found 377716 V0s, 3266316 cascades, 120432 3-body decays, 0 strange tracks. Timing: CPU: 421.99 Real: 11.212 s
[2656164:secondary-vertexing]: [13:48:06][INFO] Found 1027350 V0s, 16893633 cascades, 191060 3-body decays, 0 strange tracks. Timing: CPU: 1825.47 Real: 48.0864 s
[2656164:secondary-vertexing]: [13:48:31][INFO] Found 593069 V0s, 10863745 cascades, 152370 3-body decays, 0 strange tracks. Timing: CPU: 939.13 Real: 25.1205 s
[2656164:secondary-vertexing]: [13:48:56][INFO] Found 925954 V0s, 13250610 cascades, 240363 3-body decays, 0 strange tracks. Timing: CPU: 994.75 Real: 25.8171 s
[2656164:secondary-vertexing]: [13:49:21][INFO] Found 746878 V0s, 9346087 cascades, 214883 3-body decays, 0 strange tracks. Timing: CPU: 957.48 Real: 24.5402 s
[2656164:secondary-vertexing]: [13:50:02][INFO] Found 958278 V0s, 17381581 cascades, 258242 3-body decays, 0 strange tracks. Timing: CPU: 1601.25 Real: 41.1889 s
[2656164:secondary-vertexing]: [13:51:21][INFO] Found 1719290 V0s, 29001772 cascades, 256805 3-body decays, 0 strange tracks. Timing: CPU: 3071.3 Real: 78.4071 s
[2656164:secondary-vertexing]: [13:52:01][INFO] Found 1073049 V0s, 12315667 cascades, 260478 3-body decays, 0 strange tracks. Timing: CPU: 1599.24 Real: 40.8381 s
[2656164:secondary-vertexing]: [13:52:34][INFO] Found 933896 V0s, 10725893 cascades, 103165 3-body decays, 0 strange tracks. Timing: CPU: 1261.61 Real: 32.2567 s
[2656164:secondary-vertexing]: [13:53:07][INFO] Found 953385 V0s, 16929699 cascades, 313731 3-body decays, 0 strange tracks. Timing: CPU: 1286.16 Real: 32.9474 s
[2656164:secondary-vertexing]: [13:53:28][INFO] Found 599581 V0s, 7944977 cascades, 219717 3-body decays, 0 strange tracks. Timing: CPU: 831.85 Real: 21.485 s
[2656164:secondary-vertexing]: [13:54:13][INFO] Found 1232085 V0s, 16714986 cascades, 190138 3-body decays, 0 strange tracks. Timing: CPU: 1766.55 Real: 45.2307 s
[2656164:secondary-vertexing]: [13:55:49][INFO] Found 2354854 V0s, 46218988 cascades, 350294 3-body decays, 0 strange tracks. Timing: CPU: 3704.31 Real: 95.2266 s
[2656164:secondary-vertexing]: [13:56:15][INFO] Found 842509 V0s, 14574387 cascades, 240984 3-body decays, 0 strange tracks. Timing: CPU: 1016.42 Real: 26.1359 s
[2656164:secondary-vertexing]: [13:56:27][INFO] Found 513109 V0s, 6078201 cascades, 185116 3-body decays, 0 strange tracks. Timing: CPU: 471.93 Real: 12.1472 s
[2656164:secondary-vertexing]: [13:56:53][INFO] Found 695464 V0s, 10456996 cascades, 199036 3-body decays, 0 strange tracks. Timing: CPU: 1011.46 Real: 25.9277 s
[2656164:secondary-vertexing]: [13:57:11][INFO] Found 560730 V0s, 6027929 cascades, 56531 3-body decays, 0 strange tracks. Timing: CPU: 715.46 Real: 18.2371 s
[2656164:secondary-vertexing]: [13:58:28][INFO] Found 1662181 V0s, 29843309 cascades, 215666 3-body decays, 0 strange tracks. Timing: CPU: 3032.32 Real: 77.3317 s
[2656164:secondary-vertexing]: [13:58:53][INFO] Found 752433 V0s, 10074706 cascades, 220399 3-body decays, 0 strange tracks. Timing: CPU: 962.48 Real: 24.6174 s
[2656164:secondary-vertexing]: [13:59:28][INFO] Found 1087919 V0s, 19958251 cascades, 207563 3-body decays, 0 strange tracks. Timing: CPU: 1366.82 Real: 35.1748 s
[2656164:secondary-vertexing]: [13:59:53][INFO] Found 845448 V0s, 14173403 cascades, 183649 3-body decays, 0 strange tracks. Timing: CPU: 944.24 Real: 24.4718 s
[2656164:secondary-vertexing]: [14:00:49][INFO] Found 1065565 V0s, 24210602 cascades, 207734 3-body decays, 0 strange tracks. Timing: CPU: 2193.51 Real: 55.954 s
[2656164:secondary-vertexing]: [14:00:59][INFO] Found 457320 V0s, 4976456 cascades, 134697 3-body decays, 0 strange tracks. Timing: CPU: 391.36 Real: 10.0857 s
[2656164:secondary-vertexing]: [14:02:02][INFO] Found 1482966 V0s, 27114167 cascades, 214857 3-body decays, 0 strange tracks. Timing: CPU: 2462.25 Real: 62.972 s
[2656164:secondary-vertexing]: [14:02:22][INFO] Found 640716 V0s, 6515766 cascades, 138581 3-body decays, 0 strange tracks. Timing: CPU: 768.97 Real: 19.882 s
[2656164:secondary-vertexing]: [14:02:46][INFO] Found 870124 V0s, 10669274 cascades, 194755 3-body decays, 0 strange tracks. Timing: CPU: 931.69 Real: 24.0609 s
[2656164:secondary-vertexing]: [14:02:54][INFO] Found 395179 V0s, 3863638 cascades, 130647 3-body decays, 0 strange tracks. Timing: CPU: 319.56 Real: 8.23112 s
[2656164:secondary-vertexing]: [14:03:16][INFO] Found 764119 V0s, 9109674 cascades, 196687 3-body decays, 0 strange tracks. Timing: CPU: 854.02 Real: 21.8797 s
[2656164:secondary-vertexing]: [14:03:55][INFO] Found 1072526 V0s, 14374030 cascades, 170800 3-body decays, 0 strange tracks. Timing: CPU: 1524.31 Real: 38.992 s
[2656164:secondary-vertexing]: [14:04:16][INFO] Found 775485 V0s, 9702760 cascades, 152037 3-body decays, 0 strange tracks. Timing: CPU: 840.7 Real: 21.6295 s
[2656164:secondary-vertexing]: [14:04:21][INFO] Found 261797 V0s, 2033536 cascades, 77800 3-body decays, 0 strange tracks. Timing: CPU: 185.01 Real: 4.79786 s
[2656164:secondary-vertexing]: [14:04:35][INFO] Found 519552 V0s, 5520688 cascades, 117460 3-body decays, 0 strange tracks. Timing: CPU: 535.22 Real: 13.7109 s
[2656164:secondary-vertexing]: [14:04:43][INFO] Found 302304 V0s, 3405950 cascades, 90911 3-body decays, 0 strange tracks. Timing: CPU: 329.36 Real: 8.59948 s
[2656164:secondary-vertexing]: [14:05:20][INFO] Found 1173911 V0s, 13883810 cascades, 217732 3-body decays, 0 strange tracks. Timing: CPU: 1402.97 Real: 36.0476 s
[2656164:secondary-vertexing]: [14:06:01][INFO] Found 1041008 V0s, 19456413 cascades, 103106 3-body decays, 0 strange tracks. Timing: CPU: 1622.63 Real: 41.8946 s
[2656164:secondary-vertexing]: [14:06:25][INFO] Found 850422 V0s, 13994052 cascades, 248028 3-body decays, 0 strange tracks. Timing: CPU: 903.14 Real: 23.4953 s
[2656164:secondary-vertexing]: [14:06:50][INFO] Found 621955 V0s, 13004107 cascades, 116928 3-body decays, 0 strange tracks. Timing: CPU: 982.07 Real: 25.2074 s
```
With this PR:
```
shahoian@alicrs02:~/Downloads$ grep 'strange tracks' sv1.log 
[2686391:secondary-vertexing]: [14:56:30][INFO] Found 98059 V0s, 1 cascades, 94294 3-body decays, 0 strange tracks. Timing: CPU: 173.21 Real: 7.68342 s
[2686391:secondary-vertexing]: [14:58:44][INFO] Found 107381 V0s, 0 cascades, 93875 3-body decays, 0 strange tracks. Timing: CPU: 282.77 Real: 7.4255 s
[2686391:secondary-vertexing]: [14:59:19][INFO] Found 283986 V0s, 7 cascades, 152382 3-body decays, 0 strange tracks. Timing: CPU: 1324.85 Real: 34.9998 s
[2686391:secondary-vertexing]: [14:59:38][INFO] Found 162785 V0s, 0 cascades, 120856 3-body decays, 0 strange tracks. Timing: CPU: 708.27 Real: 19.431 s
[2686391:secondary-vertexing]: [14:59:59][INFO] Found 225825 V0s, 2 cascades, 192982 3-body decays, 0 strange tracks. Timing: CPU: 746.25 Real: 20.3846 s
[2686391:secondary-vertexing]: [15:00:19][INFO] Found 195522 V0s, 1 cascades, 169358 3-body decays, 0 strange tracks. Timing: CPU: 764.4 Real: 20.2724 s
[2686391:secondary-vertexing]: [15:00:48][INFO] Found 218031 V0s, 3 cascades, 196191 3-body decays, 0 strange tracks. Timing: CPU: 1154.2 Real: 29.1272 s
[2686391:secondary-vertexing]: [15:01:45][INFO] Found 364919 V0s, 5 cascades, 203633 3-body decays, 0 strange tracks. Timing: CPU: 2270.85 Real: 57.2427 s
[2686391:secondary-vertexing]: [15:02:15][INFO] Found 279807 V0s, 3 cascades, 205031 3-body decays, 0 strange tracks. Timing: CPU: 1164.07 Real: 29.42 s
[2686391:secondary-vertexing]: [15:02:39][INFO] Found 225297 V0s, 6 cascades, 80848 3-body decays, 0 strange tracks. Timing: CPU: 950.98 Real: 24.0352 s
[2686391:secondary-vertexing]: [15:03:02][INFO] Found 272693 V0s, 2 cascades, 248908 3-body decays, 0 strange tracks. Timing: CPU: 931.59 Real: 23.5129 s
[2686391:secondary-vertexing]: [15:03:17][INFO] Found 147172 V0s, 2 cascades, 176832 3-body decays, 0 strange tracks. Timing: CPU: 565.62 Real: 14.3294 s
[2686391:secondary-vertexing]: [15:03:48][INFO] Found 289382 V0s, 3 cascades, 151648 3-body decays, 0 strange tracks. Timing: CPU: 1255.93 Real: 31.7059 s
[2686391:secondary-vertexing]: [15:04:57][INFO] Found 482502 V0s, 17 cascades, 277565 3-body decays, 0 strange tracks. Timing: CPU: 2716 Real: 69.001 s
[2686391:secondary-vertexing]: [15:05:15][INFO] Found 234493 V0s, 3 cascades, 189174 3-body decays, 0 strange tracks. Timing: CPU: 703.35 Real: 17.7995 s
[2686391:secondary-vertexing]: [15:05:23][INFO] Found 139424 V0s, 0 cascades, 145320 3-body decays, 0 strange tracks. Timing: CPU: 310.66 Real: 7.88107 s
[2686391:secondary-vertexing]: [15:05:40][INFO] Found 186626 V0s, 3 cascades, 153884 3-body decays, 0 strange tracks. Timing: CPU: 687.46 Real: 17.4734 s
[2686391:secondary-vertexing]: [15:06:36][INFO] Found 369325 V0s, 5 cascades, 169753 3-body decays, 0 strange tracks. Timing: CPU: 2194.57 Real: 55.2309 s
[2686391:secondary-vertexing]: [15:06:49][INFO] Found 134001 V0s, 3 cascades, 44905 3-body decays, 0 strange tracks. Timing: CPU: 513.17 Real: 12.9813 s
[2686391:secondary-vertexing]: [15:07:05][INFO] Found 215048 V0s, 1 cascades, 177077 3-body decays, 0 strange tracks. Timing: CPU: 637.42 Real: 16.1258 s
[2686391:secondary-vertexing]: [15:07:28][INFO] Found 240183 V0s, 4 cascades, 164903 3-body decays, 0 strange tracks. Timing: CPU: 935.87 Real: 23.6056 s
[2686391:secondary-vertexing]: [15:08:08][INFO] Found 293346 V0s, 4 cascades, 165839 3-body decays, 0 strange tracks. Timing: CPU: 1576.2 Real: 39.7528 s
[2686391:secondary-vertexing]: [15:08:25][INFO] Found 231596 V0s, 2 cascades, 146076 3-body decays, 0 strange tracks. Timing: CPU: 652.87 Real: 16.671 s
[2686391:secondary-vertexing]: [15:08:31][INFO] Found 123069 V0s, 0 cascades, 108783 3-body decays, 0 strange tracks. Timing: CPU: 259.4 Real: 6.5798 s
[2686391:secondary-vertexing]: [15:09:15][INFO] Found 299718 V0s, 5 cascades, 161371 3-body decays, 0 strange tracks. Timing: CPU: 1738.12 Real: 43.7764 s
[2686391:secondary-vertexing]: [15:09:29][INFO] Found 142286 V0s, 2 cascades, 110073 3-body decays, 0 strange tracks. Timing: CPU: 530.35 Real: 13.6587 s
[2686391:secondary-vertexing]: [15:09:46][INFO] Found 214011 V0s, 2 cascades, 153594 3-body decays, 0 strange tracks. Timing: CPU: 664.92 Real: 16.9983 s
[2686391:secondary-vertexing]: [15:09:51][INFO] Found 118563 V0s, 1 cascades, 102649 3-body decays, 0 strange tracks. Timing: CPU: 218.11 Real: 5.55767 s
[2686391:secondary-vertexing]: [15:10:07][INFO] Found 201325 V0s, 1 cascades, 156443 3-body decays, 0 strange tracks. Timing: CPU: 606.5 Real: 15.3142 s
[2686391:secondary-vertexing]: [15:10:35][INFO] Found 224307 V0s, 2 cascades, 135556 3-body decays, 0 strange tracks. Timing: CPU: 1099.6 Real: 27.8317 s
[2686391:secondary-vertexing]: [15:10:50][INFO] Found 181359 V0s, 1 cascades, 122538 3-body decays, 0 strange tracks. Timing: CPU: 609.02 Real: 15.4165 s
[2686391:secondary-vertexing]: [15:10:53][INFO] Found 72790 V0s, 1 cascades, 61561 3-body decays, 0 strange tracks. Timing: CPU: 122.44 Real: 3.14255 s
[2686391:secondary-vertexing]: [15:11:03][INFO] Found 130822 V0s, 1 cascades, 93604 3-body decays, 0 strange tracks. Timing: CPU: 377.17 Real: 9.55567 s
[2686391:secondary-vertexing]: [15:11:08][INFO] Found 91134 V0s, 1 cascades, 70711 3-body decays, 0 strange tracks. Timing: CPU: 229.59 Real: 5.85402 s
[2686391:secondary-vertexing]: [15:11:34][INFO] Found 265628 V0s, 3 cascades, 173324 3-body decays, 0 strange tracks. Timing: CPU: 1003.28 Real: 25.4538 s
[2686391:secondary-vertexing]: [15:12:06][INFO] Found 185130 V0s, 5 cascades, 80256 3-body decays, 0 strange tracks. Timing: CPU: 1260.29 Real: 31.9637 s
[2686391:secondary-vertexing]: [15:12:21][INFO] Found 205781 V0s, 2 cascades, 201529 3-body decays, 0 strange tracks. Timing: CPU: 607.87 Real: 15.4385 s
[2686391:secondary-vertexing]: [15:12:40][INFO] Found 176657 V0s, 3 cascades, 93401 3-body decays, 0 strange tracks. Timing: CPU: 736.16 Real: 18.6499 s
```

I see that the expected Lambda width with new parametrization is 1.8MeV above 1 GeV, isn't just 4 sigma cut too tight?